pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "studyassist-intel-system"
version = "0.3.0"
description = "MVP intelligence pipeline for platforms/competitors discovery and scoring"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn[standard]>=0.35.0",
  "sqlalchemy>=2.0.36",
  "pydantic-settings>=2.6.1",
  "httpx>=0.27.2",
  "python-dotenv>=1.0.1",
  "prometheus-client>=0.21.0",
  "sentry-sdk>=2.19.2",
  "python-multipart>=0.0.20",
  "beautifulsoup4>=4.12.0",
  "PyJWT>=2.8.0"
]

[project.optional-dependencies]
# PostgreSQL backend (server/cloud deployment)
postgres = [
  "psycopg[binary]>=3.2.1",
]
# Desktop .exe build
desktop = [
  "pywebview>=5.1",
]
# Build tools
build = [
  "pyinstaller>=6.0",
  "pillow>=10.0",
]
dev = [
  "pytest>=8.3.3",
  "pytest-cov>=5.0.0",
]

[project.scripts]
studyassist-desktop = "app.launcher:main"
studyassist-server  = "app.main:app"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
